[ADD] Added Exception handlers for SedeController.java

+++ backend/src/main/java/it/unicam/cs/ids/loyaltyplatform/controller/SedeController.java
@@ -24,17 +24,32 @@ public class SedeController
 
     @PostMapping(value = "/addSede")
     public Sede addSede(@RequestBody Sede sede) {
+        validateSede(sede);
+        if (getSedeById(sede.getIdSede()) != null)
+            throw new RuntimeException("La sede inserita è già presente nel database");
         return this.sedeService.addSede(sede);
     }
 
     @DeleteMapping(value = "/deleteSedeById/{id}")
     public void deleteSedeById(@PathVariable("id") Integer id) {
+        if((getSedeById(id)) == null)
+            throw new NullPointerException("ID non valido per la cancellazione della sede");
         this.sedeService.deleteSedeById(id);
     }
 
     @PutMapping(value = "/updateSede")
     public void updateSede(@RequestBody Sede sede) {
+        validateSede(sede);
+        if (getSedeById(sede.getIdSede()) == null)
+            throw new IllegalArgumentException("Sede non presente");
         this.sedeService.updateSede(sede);
     }
 
+    private void validateSede(Sede sede) {
+        if(sede == null)
+            throw new RuntimeException("La sede inserita è nulla");
+        if (sede.getIdSede() == null || sede.getQualeAzienda() == null || sede.getVia() == null || sede.getCap() == null
+                || sede.getRegione() == null || sede.getCitta() == null || sede.getCivico() == null)
+            throw new IllegalArgumentException("Parametri non validi per la sede");
+    }
 }
